requirements.txt
# Scraper
playwright>=1.43.0
playwright-stealth>=1.0.6

# Dashboard
streamlit>=1.32.0
plotly>=5.18.0
pandas>=2.2.0

# CLI
click>=8.1.0
rich>=13.7.0

# LLM
openai>=1.12.0
python-dotenv>=1.0.0
